Allow setting class of the dialog modal div.
This follows the convention in some of the other components where
`options` is passed into the erb as options. I've chose the `:class`
here which associates with the dialog modal. This is because the trigger
does no existing classes defined and we can introduce a `:trigger_class`
in the future if needed.